+++ components/dropdown/dropdown-menu-core.jsx
@@ -40,6 +40,11 @@ export function DropdownMenuCore({ children, popoverProps, ariaProps }) {
 		}
 	}, [focusedMenuItem, focusedIndex, focusableChildList, setFocusedMenuItem]);
 
+	const handleFocusedMenuItemChange = useCallback(
+		itemIndex => () => setFocusedMenuItem(itemIndex),
+		[setFocusedMenuItem],
+	);
+
 	return (
 		<Box id={menuId} {...ariaProps}>
 			<Popover
@@ -61,6 +66,7 @@ export function DropdownMenuCore({ children, popoverProps, ariaProps }) {
 						React.isValidElement(child)
 							? React.cloneElement(child, {
 									isSelected: index === focusedMenuItem,
+									onFocusedMenuItemChange: handleFocusedMenuItemChange(index),
 							  })
 							: null,
 					)}